feat(phase1): pulse daemon + heterogeneous intelligence — 108→114 tests
Phase 1 criteria:
- Workspace compiles clean (cargo check ✓)
- CognitiveEngine routes between backends (mock/OpenAI/Anthropic) ✓
- LocalBackend with hardware auto-detection (CPU/CUDA/Metal/Vulkan) ✓
- Deterministic embeddings (384-dim semantic hash) ✓
- 6 new integration tests for backend routing and hardware detection
- Clippy warnings reduced from 35→17

Test result: 114 passing, 0 failing

+++ crates/candor-cognitive/src/real_embed.rs
@@ -32,7 +32,7 @@ pub fn deterministic_embed(text: &str, dim: usize) -> Vec<f32> {
 
         // Spread the hash across multiple dimensions
         for j in 0..8 {
-            let idx = ((hash as usize + j * 31 + i * 7) % dim as u64 as usize) as usize;
+            let idx = ((hash as usize + j * 31 + i * 7) % dim as u64 as usize);
             let bit = ((hash >> (j * 8)) & 0xFF) as f32;
             vec[idx] += (bit - 128.0) / 128.0; // Normalize to [-1, 1]
         }
